Increase request timeout; Move some components to widget folder; Fix widget mobile layout;
