+++ src/PredictionWidget/components/ErrorCard/styles.css
@@ -18,7 +18,7 @@
 }
 
 .error-card__error-title {
-    padding-bottom: 2rem;
+    padding-bottom: 2em;
 
     text-align: center;
     color: var(--main-text-color);
@@ -26,7 +26,7 @@
     font-family: var(--main-font);
     font-style: normal;
     font-weight: bold;
-    font-size: 3rem;
+    font-size: 3em;
 }
 
 .error-card__error-subtitle {
@@ -36,5 +36,5 @@
     font-family: var(--main-font);
     font-style: normal;
     font-weight: 300;
-    font-size: 1.25rem;
+    font-size: 1.25em;
 }
