+++ src/PredictionWidget/components/GeneratePredictionCard/index.js
