+++ src/PredictionWidget/components/GeneratePredictionCard/styles.css
@@ -8,7 +8,7 @@
     justify-content: center;
     align-items: center;
 
-    padding-top: 3rem;
+    padding-top: 3em;
 }
 
 .generate-prediction-card__title {
@@ -22,24 +22,23 @@
     font-family: var(--main-font);
     font-style: normal;
     font-weight: bold;
-    font-size: 4.25rem;
-    line-height: 84px;
+    font-size: 4em;
 }
 
 .generate-prediction-card__wallet {
     width: 100%;
-    padding-left: 3rem;
-    padding-right: 3rem;
-    padding-top: 1rem;
+    padding-left: 3em;
+    padding-right: 3em;
+    padding-top: 1em;
     box-sizing: border-box;
 }
 
 .generate-prediction-card__wallet-input {
     display: flex;
-    height: 4rem;
+    height: 4em;
     width: 100%;
-    padding-left: 3rem;
-    padding-right: 3rem;
+    padding-left: 3em;
+    padding-right: 3em;
     box-sizing: border-box;
 
     color: var(--main-text-color);
@@ -53,7 +52,7 @@
     font-family: var(--main-font);
     font-style: normal;
     font-weight: 500;
-    font-size: 1.75rem;
+    font-size: 1em;
     line-height: 33px;
 }
 
@@ -62,26 +61,29 @@
 }
 
 .generate-prediction-card__subtitle {
+    padding-left: 0.5em;
+    padding-right: 0.5em;
+
     color: var(--main-text-color);
     text-align: center;
 
     font-family: var(--main-font);
     font-style: normal;
     font-weight: 300;
-    font-size: 1.25rem;
+    font-size: 1.25em;
 }
 
 .generate-prediction-card__subtitle_error {
-    --error-block-height: -1rem;
+    --error-block-height: -1em;
 
     margin-top: var(--error-block-height);
 }
 
 .generate-prediction-card__error {
-    --error-block-height: 1rem;
+    --error-block-height: 1em;
     height: var(--error-block-height); /* Same as font-size */
-    padding-top: 0.5rem;
-    padding-bottom: 0.5rem;
+    padding-top: 0.5em;
+    padding-bottom: 0.5em;
 
     color: rgba(243, 30, 30, 1);
 
@@ -89,13 +91,13 @@
 }
 
 .generate-prediction-card__action {
-    padding-top: 1rem;
+    padding-top: 1em;
 }
 
 .generate-prediction-card__support-text {
-    padding-top: 1.5rem;
-    padding-left: 5rem;
-    padding-right: 5rem;
+    padding-top: 1.5em;
+    padding-left: 5em;
+    padding-right: 5em;
 
     color: var(--main-text-color);
     text-align: center;
@@ -103,31 +105,39 @@
     font-family: var(--main-font);
     font-style: italic;
     font-weight: normal;
-    font-size: 1rem;
+    font-size: 1em;
     line-height: 19px;
 }
 
 @media screen and (max-width: 425px) {
     .generate-prediction-card__wallet {
-        padding-left: 0.5rem;
-        padding-right: 0.5rem;
+        padding-left: 0.5em;
+        padding-right: 0.5em;
     }
 
     .generate-prediction-card__content {
-        padding-top: 1rem;
+        padding-top: 1em;
     }
 
     .generate-prediction-card__title {
-        font-size: 3rem;
+        font-size: 3em;
     }
 
     .generate-prediction-card__subtitle {
-        padding-left: 1rem;
-        padding-right: 1rem;
+        padding-left: 1em;
+        padding-right: 1em;
     }
 
     .generate-prediction-card__support-text {
-        padding-left: 1rem;
-        padding-right: 1rem;
+        padding-left: 1em;
+        padding-right: 1em;
+    }
+
+    .generate-prediction-card__wallet-input {
+        height: 3em;
+
+        padding-left: 1.5em;
+        padding-right: 1.5em;
+        font-size: 1.5em;
     }
 }
