+++ src/PredictionWidget/components/PredictionCard/index.js
