+++ src/PredictionWidget/styles.css
@@ -2,8 +2,32 @@
     position: relative;
     height: 100%;
     min-height: 530px;
+
+    font-size: 16px;
 }
 
 .prediction-widget__form {
     height: 100%;
+}
+
+@media screen and (max-width: 768px) {
+    .prediction-widget {
+        font-size: 14px;
+    }
+}
+
+@media screen and (max-width: 425px) {
+    .prediction-widget {
+        font-size: 14px;
+    }
+
+    .card__header {
+        padding-left: 1em;
+        padding-right: 1em;
+    }
+
+    .card__header-title {
+        padding-left: 0.5em;
+        padding-right: 0.5em;
+    }
 }
