+++ src/components/Card/styles.css
@@ -2,7 +2,7 @@
     display: flex;
     flex-flow: column nowrap;
 
-    border-radius: 1.5rem;
+    border-radius: 1.5em;
     background-color: rgba(181, 181, 181, 0.1);
     backdrop-filter: blur(10px);
 }
@@ -12,9 +12,9 @@
     justify-content: space-between;
     align-items: center;
 
-    padding-top: 1.5rem;
-    padding-left: 2rem;
-    padding-right: 2rem;
+    padding-top: 1.5em;
+    padding-left: 2em;
+    padding-right: 2em;
 }
 
 .card__header-info {
@@ -23,22 +23,22 @@
 }
 
 .card__header-title {
-    padding-left: 1.5rem;
-    padding-right: 1.5rem;
+    padding-left: 1.5em;
+    padding-right: 1.5em;
     text-align: center;
     color: var(--main-text-color);
 
     font-family: var(--main-font);
     font-style: normal;
     font-weight: 800;
-    font-size: 1.5rem;
-    line-height: 1.5rem;
+    font-size: 1.5em;
+    line-height: 1.5em;
 }
 
 .card__content {
-    padding-top: 2rem;
+    padding-top: 2em;
 }
 
 .card__footer {
-    padding-bottom: 1.5rem;
+    padding-bottom: 1.5em;
 }
