+++ src/components/CardSpinner/styles.css
@@ -7,7 +7,7 @@
     justify-content: center;
     align-items: center;
 
-    border-radius: 1.5rem;
+    border-radius: 1.5em;
 
     color: var(--main-text-color);
     background-color: rgba(255, 255, 255, 0.3);
