+++ src/components/SimpleCard/styles.css
@@ -2,7 +2,7 @@
     background-color: rgba(18, 18, 18, 0.15);
     border: 2px solid rgba(255, 255, 255, 0.5);
     box-sizing: border-box;
-    backdrop-filter: blur(3rem);
+    backdrop-filter: blur(3em);
 
     border-radius: var(--card-border-radius);
 }
