+++ src/notFound/NotFound.js
@@ -3,7 +3,7 @@ import { Helmet } from "react-helmet";
 import { Link } from "react-router-dom";
 import Background from "../components/Background";
 import Header from "../components/Header";
-import Card from "./components/Card/Card";
+import Card from "../components/Card";
 import { Button } from "../components/Button";
 
 import "./styles.css";
